refactor: rename crates for consistent strided- prefix
- stridedview → strided-view (crate: strided-view)
- strided → strided-kernel (crate: strided-kernel)
- strided-einsum2 unchanged

Also rename internal module strided_view.rs → view.rs to avoid
name collision between the crate identifier and module name.
