+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["stridedview", "strided", "strided-einsum2"]
+members = ["strided-view", "strided-kernel", "strided-einsum2"]
 resolver = "2"
