+++ strided-kernel/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "strided"
+name = "strided-kernel"
 version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
@@ -9,7 +9,7 @@ description = "Cache-optimized kernels for strided multidimensional array operat
 publish = false
 
 [dependencies]
-stridedview = { path = "../stridedview" }
+strided-view = { path = "../strided-view" }
 num-traits = "0.2"
 num-complex = "0.4"
 rayon = { version = "1.10", optional = true }
