+++ strided-kernel/benches/run_scaling.sh
