+++ strided-kernel/benches/run_threaded.sh
