+++ strided-kernel/src/fuse.rs
@@ -173,7 +173,7 @@ pub fn compute_costs(all_strides: &[&[isize]]) -> Vec<isize> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use stridedview::auxiliary::index_order;
+    use strided_view::auxiliary::index_order;
 
     #[test]
     fn test_fuse_dims_contiguous() {
