+++ strided-view/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "stridedview"
+name = "strided-view"
 version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
