+++ strided-view/src/auxiliary.rs
